% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{concepts}
\alias{concepts}
\title{Default Concept Vocabulary}
\format{
A data frame with 80392 rows and 5 variables:
}
\usage{
concepts
}
\description{
A loadable concept vocabulary for OMOP Genomic
}
\keyword{datasets}
