% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eta}
\alias{eta}
\title{ETA of completion'}
\usage{
eta(x, max, start)
}
\arguments{
\item{x}{Current progress}

\item{max}{Final progress value}

\item{start}{A Sys.time() call at the beginning of processing}
}
\description{
ETA of completion'
}
