% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic.R
\name{findReference}
\alias{findReference}
\title{Attempt to uncover the underlying reference genome used to generate a given VCF file}
\usage{
findReference(vcf)
}
\arguments{
\item{vcf}{A vcf object read by loadVCF()}
}
\value{
A string indicating the reference genome used
}
\description{
Attempt to uncover the underlying reference genome used to generate a given VCF file
}
