% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHgvsg.R
\name{generateHGVSG}
\alias{generateHGVSG}
\title{Process a single vcfR object to return a dataframe containing hgvsg notation}
\usage{
generateHGVSG(vcf.df, ref.df)
}
\arguments{
\item{vcf.df}{Input vcf.df object}

\item{ref.df}{A ref.df object containing chrom specifications}
}
\value{
A dataframe containing hgvsg notation for all variants
}
\description{
Process a single vcfR object to return a dataframe containing hgvsg notation
}
