% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHgvsg.R
\name{hgvsgConvert}
\alias{hgvsgConvert}
\title{Generate a hgvsg string}
\usage{
hgvsgConvert(row, ref.df)
}
\arguments{
\item{row}{A vcf.df row}

\item{ref.df}{A ref.df object containing chrom specifications}
}
\value{
A single hgvsg string
}
\description{
Generate a hgvsg string
}
