% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHgvsg.R
\name{hgvsgConvert_cBioPortal}
\alias{hgvsgConvert_cBioPortal}
\title{Generate a hgvsg string from cBioPortal format}
\usage{
hgvsgConvert_cBioPortal(row, ref.df)
}
\arguments{
\item{row}{A mut_vcf row}

\item{ref.df}{A ref.df object containing chrom specifications}
}
\value{
A single hgvsg string
}
\description{
Generate a hgvsg string from cBioPortal format
}
