% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputOutput.R
\name{loadConcepts}
\alias{loadConcepts}
\title{Generate a set of processed alignments given a stringDF dataframe}
\usage{
loadConcepts()
}
\value{
A set of concept synonyms for the OMOP Genomic dataset, derived from ATHENA
}
\description{
Generate a set of processed alignments given a stringDF dataframe
}
