% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputOutput.R
\name{loadConcepts_fusions}
\alias{loadConcepts_fusions}
\title{Load an extended set of protein and transcript KOIOS concepts}
\usage{
loadConcepts_fusions()
}
\value{
A set of concept synonyms for the OMOP Genomic dataset, derived from ATHENA
}
\description{
Load an extended set of protein and transcript KOIOS concepts
}
