% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputOutput.R
\name{loadReference}
\alias{loadReference}
\title{Generate a set of reference maps between chromosome IDs}
\usage{
loadReference(ref = ref)
}
\arguments{
\item{ref}{The chosen reference genome, hg18/hg19/hg38}
}
\value{
A map of synonyms between chromosome IDs for a given reference genome
}
\description{
Generate a set of reference maps between chromosome IDs
}
