% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputOutput.R
\name{loadVCF}
\alias{loadVCF}
\title{Load a single .vcf file or a set of .vcf's located in a single directory}
\usage{
loadVCF(userVCF)
}
\arguments{
\item{userVCF}{A user-defined .vcf file or a directory containing several .vcf objects}
}
\value{
Either a single vcfR object or a list of vcfR objects
}
\description{
Load a single .vcf file or a set of .vcf's located in a single directory
}
