% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHgvsg.R
\name{processcBioPortal}
\alias{processcBioPortal}
\title{Process a single cBioPortal data object to return a dataframe containing hgvsg notation}
\usage{
processcBioPortal(mutations)
}
\arguments{
\item{mutations}{Input mutations data object}
}
\value{
A dataframe containing hgvsg notation for all variants
}
\description{
Process a single cBioPortal data object to return a dataframe containing hgvsg notation
}
