% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{concepts}
\alias{concepts}
\title{Default Concept Vocabulary}
\format{
A data frame containing genomic variants
\describe{
  \item{concept_id}{OMOP Concept ID}
  \item{concept_name}{OMOP Concept Name}
  \item{concept_synonym_name}{OMOP Concept Synonyms in HGVSG format}
  \item{ver}{Specific version of reference genome}
}
}
\usage{
concepts
}
\description{
A loadable concept vocabulary for OMOP Genomic
}
\keyword{datasets}
