% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputOutput.R
\name{loadConcepts}
\alias{loadConcepts}
\title{Load a set of genomic KOIOS concepts}
\usage{
loadConcepts()
}
\value{
A set of concept synonyms for the OMOP Genomic dataset, derived from ATHENA
}
\description{
Load a set of genomic KOIOS concepts
}
