% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processClingen.R
\name{processClinGen}
\alias{processClinGen}
\title{Process a vcf.df dataframe to return clingen data for all variants}
\usage{
processClinGen(vcf.df, ref, generateAll = FALSE, progressBar = TRUE)
}
\arguments{
\item{vcf.df}{Input vcf.df object}

\item{ref}{A string indicating the reference genome used
Must be one of: hg18
                hg19
                hg38}

\item{generateAll}{A toggle indicating whether all transcript alleles, as well as
all genomic alleles, should be returned}

\item{progressBar}{A toggle indicating whether or not a progress bar will be displayed}
}
\value{
A dataframe containing data derived from the relevant ClinGen URL
for each allele found corresponding to a vcf row
}
\description{
Process a vcf.df dataframe to return clingen data for all variants
}
